Module ic_crypto_internal_threshold_sig_bls12381::api

External API for Threshold Signatures

The threshold signature scheme has the following participants:

Signatories are actors who hold a threshold key and can sign with that key.

Dealers are actors who generate keys that are then distributed to signatories.

Verifiers are actors who can verify that signatures are valid.

The key properties of this threshold signature scheme:

The public output of key generation is a set of PublicCoefficients.

Individual signatures can be verified against the public coefficients.

Individual signatures can be combined. If there are enough individual signatures, and they are all valid, the combined signature is guaranteed to be valid and can be validated against the PublicCoefficients.

The minimum number of individual signatures needed to make a valid combined signature is called the threshold.

A valid combined signature is unique and the same, regardless of which threshold signatures were combined.

The combined signature has constant size, regardless of the number of signatures or the threshold. In this implementation that constant size is the same as a single individual signature.

If some individual signatures are invalid, the combined signature is likely to be invalid. Individual signatures SHOULD be checked before combining.

Implementation notes:

Potential signatories are stored in a list. It is important to preserve the order of signatories, especially in the case of distributed key generation where multiple keys may be dealt for each signatory.

Given that signatories may become ineligible the standard form for storing signatory information is <Vec<Option<T>>, where:

the index in the vector corresponds to the signatory index

the Option is set to None for ineligible, discredited or otherwise non-participating signatories.

The external API methods are equivalent to internal methods but use opaque data types. The external API parses requests to the internal types, calls the corresponding internal methods and serialises the responses.

Modules

dkg_errors
Error types used in the public DKG API.

ni_dkg_errors
Error types used by non-interactive DKG.

threshold_sign_error
Error encountered during threshold signing.

Functions

combine_signatures
Combines individual signatures.

combined_public_key
Extracts the combined public key from the PublicCoefficients.

combined_signature_and_public_key
Generates new keys for threshold signatories and a signature using the new keys.

individual_public_key
Derives the public key of one signatory from the public_coefficients.

keygen
Generates keys for threshold signatories.

public_key_from_der
Parses a PublicKeyBytes from its DER-encoded form.

public_key_to_der
Converts public key bytes into its DER-encoded form.

sign_message
Creates an individual signature.

verify_combined_signature
Verifies that a combined signature is valid.

verify_individual_signature
Verifies that an individual signature is valid.