Module ic_crypto_internal_threshold_sig_bls12381::api::dkg_errors

Error types used in the public DKG API.

This defines two types of errors:

Individual error conditions;

Enumerations of all the error conditions that a method can return.

Structs

InvalidArgumentError
Occurs if an argument is invalid.

KeyNotFoundError
The secret key was not found in the secret key store.

MalformedDataError
Malformed X for other types of X.

MalformedPopError
Proof of possession could not be parsed.

MalformedPublicKeyError
Occurs if a public key is malformed.

MalformedSecretKeyError
Cognate to CryptoError::MalformedSecretKey

SizeError
A size is unsupported by this machine; this is not a protocol error as other machines may be able to complete this instruction successfully.

Enums

DkgCreateDealingError
Dealing of shares during DKG failed.

DkgCreateEphemeralError
Creation of an ephemeral key failed.

DkgCreateReshareDealingError
Creation of a DKG resharing failed.

DkgCreateReshareTranscriptError
Creation of a DKG transcript from a resharing failed.

DkgCreateResponseError
Creation of a DKG response failed.

DkgCreateTranscriptError
Creation of a DKG transcript failed.

DkgLoadPrivateKeyError
Loading of a private key from a DKG transcript failed.

DkgVerifyDealingError
Verification of a dealing during DKG failed.

DkgVerifyEphemeralError
Verification of an ephemeral key failed.

DkgVerifyReshareDealingError
Verification of a DKG resharing failed.

DkgVerifyResponseError
Verification of a DKG response failed.