Function ic_crypto_internal_threshold_sig_bls12381::api::combine_signatures

pub fn combine_signatures(
    signatures: &[Option<IndividualSignatureBytes>], 
    threshold: NumberOfNodes
) -> CryptoResult<CombinedSignatureBytes>

Combines individual signatures.

Note: Individual signatures SHOULD be checked for validity before being combined.

Arguments

signatures is an array that contains the signature of the ith signatory in signatures[i]. Missing signatures are represented by signatures[i]=None. It is important that the signatures have the same indices as the corresponding signatories had during key generation.

threshold is the minimum number of signatures needed.

Panics

This method is not expected to panic.

Errors

If signatures cannot be parsed, this will return an error. Given that the caller SHOULD verify individual signatures before combining them this SHOULD never occur.

If there are fewer than threshold signatures, this will return an error.