Function ic_crypto_internal_threshold_sig_bls12381::api::combined_public_key

pub fn combined_public_key(
    public_coefficients: &PublicCoefficientsBytes
) -> CryptoResult<PublicKeyBytes>

Extracts the combined public key from the PublicCoefficients.

The combined public key is used to verify combined threshold signatures.

Arguments

public_coefficients is the public output of a key generation.

Panics

This method is not expected to panic.

Errors

This method is not expected to return an error.