Function ic_crypto_internal_threshold_sig_bls12381::api::combined_signature_and_public_key

pub fn combined_signature_and_public_key(
    seed: Randomness, 
    group_size: usize, 
    threshold: NumberOfNodes, 
    message: &[u8]
) -> (CombinedSignatureBytes, PublicKeyBytes)

Generates new keys for threshold signatories and a signature using the new keys.

This is only used for testing (cf. CRP-622 and ic-crypto-internal-csp::imported_utilities:: combined_threshold_signature_and_public_key).

Arguments

seed is a random input. It must be treated as a secret.

message is the bytes to be signed.

group_size is the number of signatories.

Panics

If any one of key generation, signing, signature combination, or public key combination fails.