Function ic_crypto_internal_threshold_sig_bls12381::api::keygen

pub fn keygen(
    seed: Randomness, 
    threshold: NumberOfNodes, 
    signatory_eligibilty: &[bool]
) -> CryptoResult<(PublicCoefficientsBytes, Vec<Option<SecretKeyBytes>>)>

Generates keys for threshold signatories.

Arguments

seed is a random input. It must be treated as a secret.

threshold is the minimum number of signatures that can be combined to make a valid threshold signature.

signatory_eligibility is a boolean indicating, for each signatory, whether they should receive a key. The ith signatory should receive a key if and only if signatory_eligibilty[i]==true.

Returns

PublicCoefficients can be used by the caller to verify signatures.

Vec<Option<SecretKeyBytes>> contains secret keys. The vector has the same length as the input signatory_eligibility and the i'th entry contains a secret key if and only if signatory_eligibility[i] is true.

Panics

This method is not expected to panic.

Errors

If threshold > signatory_eligibility.len() then it is impossible for the signatories to create a valid combined signature, so this is treated as an error.