Function ic_crypto_internal_threshold_sig_bls12381::api::verify_individual_signature

pub fn verify_individual_signature(
    message: &[u8], 
    signature: IndividualSignatureBytes, 
    public_key: PublicKeyBytes
) -> CryptoResult<()>

Verifies that an individual signature is valid.

Usage note: Key generation by any compatible method generates PublicCoefficients and fixes an index for each signatory. An individual signatory's public key can be derived from the PublicCoefficients and the signatory's index with individual_public_key(..).

Arguments:

message is the bytes that have been signed. Note: This may be changed to a [u8;32] and renamed to hash pending discussion. See TODO: DFN-1430

signature is the individual signature to be verified.

public_key is the individual public key of the signatory.

Panics

This method is not expected to panic.

Errors

If signature or public_key cannot be parsed, this will return an error.