Module ic_crypto_internal_threshold_sig_bls12381::crypto

Threshold signatures with a simple dealing mechanism.

Functions

combine_signatures
Combines signature shares (i.e. evaluates the signature at x=0).

combined_public_key
Computes the public key used to verify combined signatures.

hash_message_to_g1
Hashes msg to a point in G1.

individual_public_key
Computes the public key of the index'th share from the given public coefficients of the polynomial.

keygen
Generates keys for a (t,n)-threshold signature scheme.

keygen_with_secret
Generates keys for a (t,n)-threshold signature scheme, using an existing secret key.

public_key_from_secret_key
Computes the public equivalent of a secret key.

secret_key_is_consistent
Verifies that a threshold secret key is consistent with the given public coefficients.

sign_message
Signs a message with the given secret key.

verify_combined_sig
Verifies a combined signature against the provided public key.

verify_individual_sig
Verifies an individual signature against the provided public key.

x_for_index
Yields the polynomial-evaluation point x given the index of the corresponding share.