Function ic_crypto_internal_threshold_sig_bls12381::crypto::combine_signatures

pub fn combine_signatures(
    signatures: &[Option<G1>], 
    threshold: NumberOfNodes
) -> CryptoResult<G1>

Combines signature shares (i.e. evaluates the signature at x=0).

Note: The threshold signatories are indexed from 0 to num_signatories-1. The index of each signatory defines the x-value at which the the signature is computed, so is needed along with the signature for the signature to be useful. Signatures are given in the same order as the signatories. Missing signatures are represented by None.

Errors

CryptoError::InvalidArgument if the given signature shares are lower than the given threshold.