Function ic_crypto_internal_threshold_sig_bls12381::crypto::keygen_with_secret

pub fn keygen_with_secret(
    seed: Randomness, 
    threshold: NumberOfNodes, 
    share_indices: &[bool], 
    secret: &Fr
) -> Result<(PublicCoefficients, Vec<Option<Fr>>), InvalidArgumentError>

Generates keys for a (t,n)-threshold signature scheme, using an existing secret key.

This method is identical to keygen(..) except that the threshold secret key is specified (i.e. the constant term of the randomly-generated polynomial is set to secret).

Arguments

seed - randomness used to seed the PRNG for generating the polynomial. It must be treated as a secret.

threshold - the minimum number of individual signatures that can be combined into a valid threshold signature.

share_indices - a vector with one entry per receiver. The entry is true iff the receiver is eligible to receive a threshold key.

secret - an existing secret key, which is to be shared.

Errors

This returns an error if:

The number of share indices is too large to be stored as type NumberOfNodes.

The number of eligible receivers is below the threshold; under these circumstances the receivers could never generate a valid threshold key.

The threshold is 0.