Function ic_crypto_internal_threshold_sig_bls12381::crypto::secret_key_is_consistent

pub fn secret_key_is_consistent(
    secret: Fr, 
    public_coefficients: &PublicCoefficients, 
    index: NodeIndex
) -> bool

Verifies that a threshold secret key is consistent with the given public coefficients.

Returns

true iff the threshold secret key is consistent, i.e. if the public key corresponding to the secret key is on the polynomial defined by the public coefficients.