Function ic_crypto_internal_threshold_sig_bls12381::crypto::sign_message

pub fn sign_message(message: &[u8], secret_key: &Fr) -> G1

Signs a message with the given secret key.

Note: As the whole message needs to be provided, this is unsuitable for signing large chunks of data or streaming data. For large chunks of data it is better to hash the data separately and provide the digest to sign_hash(digest: [u8: 32], secret_key: &SecretKey) // unimplemented.