Module ic_crypto_internal_threshold_sig_bls12381::dkg::secp256k1

(deprecated) Distributed Key Generation using secp256k1.

Modules are used in this order:

ephemeral_key for key generation

dealing (uses dh)

response (uses complaint)

transcript

Modules

types
(deprecated) Interactive Distributed Key Generation (DKG) Types

Functions

compute_private_key
Computes a receiver's threshold secret key.

create_dealing
Creates a new dealing: Generates threshold keys, encrypts each secret key.

create_ephemeral
Generates an ephemeral key pair, with pop.

create_resharing_dealing
Creates a resharing dealing.

create_resharing_transcript
Combines all valid resharing dealings into a short transcript.

create_response
Checks whether my share is correct in all dealings and responds with any complaints.

create_transcript
Combines all valid dealings into a short transcript.

verify_dealing
Verifies the public parameters of the dealing (degree and group size)

verify_ephemeral
Verifies an ephemeral key (e.g. for cases where subgroup checks are necessary)

verify_resharing_dealing
Verifies the public parameters of the dealing.

verify_response
Verifies the response against all dealings. This is used before including the response in a block.