Function ic_crypto_internal_threshold_sig_bls12381::dkg::secp256k1::compute_private_key

pub fn compute_private_key(
    receiver_secret_key_bytes: EphemeralSecretKeyBytes, 
    transcript: &CLibTranscriptBytes, 
    dkg_id: IDkgId
) -> Result<Option<ThresholdSecretKeyBytes>, DkgLoadPrivateKeyError>

Computes a receiver's threshold secret key.

Arguments

receiver_secret_key - the receiver's secret key, used to decrypt the threshold key.

transcript - contains all the threshold keys, including the public coefficients and (encrypted) secret keys.

dkg_id - the key distribution ID, needed when decrypting the threshold key as the IDkgId is not in the transcript.

Returns

Returns the threshold secret key, if one was issued for this node, and None otherwise.

Panics

This method is not expected to panic.

Errors

This method MUST return an error if:

any of the arguments is malformed.

the computed secret threshold key is not consistent with the public coefficients.