Function ic_crypto_internal_threshold_sig_bls12381::dkg::secp256k1::create_ephemeral

pub fn create_ephemeral<R: Rng + CryptoRng>(
    rng: &mut R, 
    dkg_id: IDkgId, 
    sender: &[u8]
) -> (EphemeralSecretKeyBytes, EphemeralPublicKeyBytes, EphemeralPopBytes)

Generates an ephemeral key pair, with pop.

Arguments

rng - a cryptographically secure random number generator.

dkg_id - the DKG this ephemeral key is to be used for.

sender - a name to identify the current node.

Returns

EphemeralSecretKeyBytes - a key that needs to be kept secret by the current node and used for the DKG protocol. It is not needed once the threshold keys have been derived.

EphemeralPublicKeyBytes and EphemeralPopBytes - the corresponding public key and proof of possession.

Panics

This method is not expected to panic.