Function ic_crypto_internal_threshold_sig_bls12381::dkg::secp256k1::create_resharing_dealing

pub fn create_resharing_dealing(
    seed: Randomness, 
    dealer_ephemeral_secret_key_bytes: EphemeralSecretKeyBytes, 
    dkg_id: IDkgId, 
    threshold: NumberOfNodes, 
    receiver_keys: &[Option<(EphemeralPublicKeyBytes, EphemeralPopBytes)>], 
    reshared_threshold_secret_key_bytes: ThresholdSecretKeyBytes
) -> Result<CLibDealingBytes, DkgCreateReshareDealingError>

Creates a resharing dealing.

The special property of a resharing key generation is that the dealer predefines the combined threshold secret key.

Arguments

seed is a secret random number used to generate the threshold keys.

dealer_ephemeral_secret_key_bytes is used to encrypt shares.

threshold is the minimum number of receivers needed to sign a valid threshold signature.

dkg_id is the identifier of the DKG instance this dealing is used for.

receiver_keys is the ephemeral public keys of the receivers, used to encrypt their shares. Given that some receivers may be ineligible this takes the standard form <Vec<Option<EphemeralPublicKey>>, where:

the index in the vector corresponds to the receiver index

the Option is set to None for ineligible, discredited or otherwise non-participating receivers.

reshared_threshold_secret_key is the dealer's preceding threshold secret key; it is shared in this dealing.

Panics

Threshold keys should be generated for every participating receiver. This method panics if there is a mismatch, i.e.:

there is a receiver with an ephemeral key for which no threshold key is generated, OR

there is a receiver with no ephemeral key for which a threshold key is generated.

Errors

The number of eligible receivers should be equal to or greater than the threshold; otherwise this MUST return an error. Note: This is checked in the threshold library by keygen(..).

The public keys must be well formed, otherwise this MUST return an error.

reshared_secret_key must be well formed, otherwise this MUST return an error.

Returns

The returned data structure contains the threshold public coefficients and encrypted threshold secret keys. Each encrypted secret key can be decrypted by the corresponding receiver.