Function ic_crypto_internal_threshold_sig_bls12381::dkg::secp256k1::create_resharing_transcript

pub fn create_resharing_transcript(
    threshold: NumberOfNodes, 
    verified_dealings: &BTreeMap<EphemeralPublicKeyBytes, CLibDealingBytes>, 
    verified_responses: &[Option<CLibVerifiedResponseBytes>], 
    dealer_keys: &[Option<(EphemeralPublicKeyBytes, EphemeralPopBytes)>], 
    resharing_public_coefficients: &PublicCoefficientsBytes
) -> Result<CLibTranscriptBytes, DkgCreateReshareTranscriptError>

Combines all valid resharing dealings into a short transcript.

Note: This is similar to the normal create_transcript except for the mechanics of combining dealings.

Arguments

threshold is the minimum number of signatures needed for a valid threshold signature.

verified_dealings are the dealings that passed initial verification. Note: This takes the standard form for per-dealer data: A map from the dealer's public key to the dealing.

verified_responses are the responses that passed initial verification. Note: this takes the standard form for per-receiver data: A vector of options which are Some for valid receivers and None for disqualified or non-participating receivers.

dealer_keys are the ephemeral keys of the dealers. The position of the dealer in the representation should match their position as a signatory in the preceding threshold key.

resharing_public_coefficients are the public coefficients of the preceding threshold key.

Prerequisites

We assume that all dealings are consistent with the threshold. This SHOULD have been confirmed when validating the dealings.

Panics

This method is not expected to panic.

Errors

This method MUST return an error if:

Any arguments are malformed.

The number of dealings is not at least threshold.

The number of responses is not at least 2 * threshold - 1.

The number of valid dealings is not at least the threshold of the preceding threshold key.