Function ic_crypto_internal_threshold_sig_bls12381::dkg::secp256k1::create_response

pub fn create_response(
    seed: Randomness, 
    receiver_secret_key_bytes: &EphemeralSecretKeyBytes, 
    dkg_id: IDkgId, 
    verified_dealings: &BTreeMap<EphemeralPublicKeyBytes, CLibDealingBytes>, 
    receiver_index: NodeIndex
) -> Result<CLibResponseBytes, DkgCreateResponseError>

Checks whether my share is correct in all dealings and responds with any complaints.

Arguments

seed - A random seed used to generate complaints.

receiver_secret_key_bytes - the current node is the receiver; this is the current node's secret ephemeral key used to decrypt the threshold key share.

dkg_id - identifier for the key generation

verified_dealings - for each qualified dealer, the dealer's public key and the key share for the current node.

receiver_index - each receiver has an index; this is the index for the current node.

Panics

This method is not expected to panic.

Errors

This method MUST return an error if:

one of the keys is malformed. It is the caller's responsibility to verify that the data is well formed before calling this method.