Function ic_crypto_internal_threshold_sig_bls12381::dkg::secp256k1::verify_resharing_dealing

pub fn verify_resharing_dealing(
    threshold: NumberOfNodes, 
    receiver_keys: &[Option<(EphemeralPublicKeyBytes, EphemeralPopBytes)>], 
    dealing: CLibDealingBytes, 
    dealer_index: NodeIndex, 
    resharing_public_coefficients: PublicCoefficientsBytes
) -> Result<(), DkgVerifyReshareDealingError>

Verifies the public parameters of the dealing.

This performs the same checks as verify_dealing but additionally verifies that the new threshold public key is the same as the individual threshold public key of the dealer in the reshared parameters.

Arguments

threshold is the minimum number of receivers needed to sign a valid threshold signature.

receiver_keys is the ephemeral public keys of the receivers, used to encrypt their shares. Given that some receivers may be ineligible this takes the standard form <Vec<Option<EphemeralPublicKey>>, where:

the index in the vector corresponds to the receiver index

the Option is set to None for ineligible, discredited or otherwise non-participating receivers.

dealing is the dealing being verified.

dealer_index is the index of the dealer as a receiver in the preceding threshold key.

resharing_public_coefficients are the public coefficients of the preceding threshold key.

Panics

This method is not expected to panic

Errors

This method returns an error if:

the public coefficients or receiver data are malformed;

the number of public coefficients does not match the threshold;

the length of dealing.receiver_data is not the same as the length of receiver_keys;

the set of receivers with a share does not match the set with public keys, i.e. dealing.receiver_data[i] is Some where receiver_keys[i] is None or vice versa.

the new threshold public key does not match the previous individual public key of the dealer.