Function ic_crypto_internal_threshold_sig_bls12381::dkg::secp256k1::verify_response

pub fn verify_response(
    dkg_id: IDkgId, 
    verified_dealings: &BTreeMap<EphemeralPublicKeyBytes, CLibDealingBytes>, 
    receiver_index: NodeIndex, 
    receiver_public_key_bytes: EphemeralPublicKeyBytes, 
    response: &CLibResponseBytes
) -> Result<(), DkgVerifyResponseError>

Verifies the response against all dealings. This is used before including the response in a block.

Arguments

dkg_id - identifier for the key generation

verified_dealings - for each qualified dealer, their dealing.

receiver_index - each receiver has an index; this is the index for the responding receiver.

receiver_key - the receiver's public key, used to verify that the receiver's complaints are legitimate.

response - the response being verified.

Panics

This method is not expected to panic.

Errors

This method MUST return an error if:

the length of verified_dealings is not the same as that of response.

a complaint in the response is invalid.