Module ic_crypto_internal_threshold_sig_bls12381::dkg::secp256k1::types

(deprecated) Interactive Distributed Key Generation (DKG) Types

The types in this file correspond to the types used in the spec, including ephemeral (secp256k1) keys, dealings, complaints and transcripts. Please see the specification for more details.

Modules

serialisation

Structs

CLibComplaint

CLibComplaintBytes

CLibDealing

CLibDealingBytes

CLibResponseBytes

CLibTranscriptBytes
This is the long term public record of a DKG.

CLibVerifiedDealing

CLibVerifiedDealingBytes

CLibVerifiedResponseBytes

EncryptedShareBytes

EphemeralKeySetBytes

EphemeralPop

EphemeralPopBytes

EphemeralPublicKey

EphemeralPublicKeyBytes
The public key as byte array.

EphemeralSecretKey

EphemeralSecretKeyBytes

Constants

SECP256K1_SECRET_KEY_ONE

SECP256K1_PUBLIC_KEY_ONE

Type Definitions

EncryptedShare