Module ic_crypto_internal_threshold_sig_bls12381::ni_dkg::groth20_bls12_381

Non-interactive Distributed Key Generation using Groth20 with BLS12-381.

Modules

types
Types for the Groth20-BLS12-381 implementation of Non-interactive Distributed Key Generation.

Functions

compute_threshold_signing_key
Computes a participant's threshold signing key from the DKG transcript.

create_dealing
Creates a new dealing, i.e. generates threshold keys.

create_forward_secure_key_pair
Generates a forward secure key pair.

create_resharing_transcript
Creates an NiDKG transcript with the same public key as an existing threshold key.

create_transcript
Creates an NiDKG transcript.

public_key_into_miracl
Parses a standard form public key into the miracl-based type.

update_forward_secure_epoch
Forgets keys before the given epoch.

verify_dealing
Verifies a dealing.

verify_forward_secure_key
Verifies that a public key is a point on the curve and that the proof of possession holds.

verify_resharing_dealing
Verifies a dealing created to reshare an existing secret.