Function ic_crypto_internal_threshold_sig_bls12381::ni_dkg::groth20_bls12_381::compute_threshold_signing_key

pub fn compute_threshold_signing_key(
    transcript: &Transcript, 
    receiver_index: NodeIndex, 
    fs_secret_key: &FsEncryptionSecretKey, 
    epoch: Epoch
) -> Result<SecretKeyBytes, CspDkgLoadPrivateKeyError>

Computes a participant's threshold signing key from the DKG transcript.

Arguments

transcript - The transcript of the distributed key generation ceremony.

receiver_index - The index of the receiver whose signing key is computed.

fs_secret_key - The forward-secure decryption key of the given receiver_index.

epoch - The forward-secure decryption epoch to use.

Errors

CspDkgLoadPrivateKeyError::InvalidTranscriptError if decryption of the share fails, or the share is invalid.

Panics

If Langrange interpolation fails (because of duplicate x-coordinates).

Transcript serialization fails (during creation of an error message).