Function ic_crypto_internal_threshold_sig_bls12381::ni_dkg::groth20_bls12_381::create_dealing

pub fn create_dealing(
    keygen_seed: Randomness, 
    encryption_seed: Randomness, 
    threshold: NumberOfNodes, 
    receiver_keys: &BTreeMap<NodeIndex, FsEncryptionPublicKey>, 
    epoch: Epoch, 
    dealer_index: NodeIndex, 
    resharing_secret: Option<ThresholdSecretKeyBytes>
) -> Result<Dealing, CspDkgCreateReshareDealingError>

Creates a new dealing, i.e. generates threshold keys.

Arguments

keygen_seed - randomness used to seed the PRNG for generating the keys/shares. It must be treated as a secret.

encryption_seed - randomness used to seed the PRNG for encrypting the shares and proving their correctness. It must be treated as a secret.

threshold - the minimum number of individual signatures that can be combined into a valid threshold signature.

receiver_keys - forward-secure encryption public keys of the receivers.

epoch - forward-secure epoch under which the shares are encrypted.

dealer_index - index of the dealer.

resharing_secret - existing secret key if resharing, or None if not.

Errors

CspDkgCreateReshareDealingError::SizeError if the length of receiver_keys isn't supported.

CspDkgCreateReshareDealingError::InvalidThresholdError if the threshold is either zero or greater than the number of receivers.

CspDkgCreateReshareDealingError::MalformedReshareSecretKeyError if resharing_key is malformed.

CspDkgCreateReshareDealingError::InvalidThresholdError if key generation fails.

CspDkgCreateDealingError::MalformedFsPublicKeyError if one of the receiver_keys is malformed.

CspDkgCreateReshareDealingError::MisnumberedReceiverError if the receiver indices are not 0..num_receivers-1 inclusive.

Panics:

If key generation produces key shares with non-contiguous indices.

If key generation produces key shares that don't match the given receiver_keys.

If the key generation produces public coefficients that are malformed.