Function ic_crypto_internal_threshold_sig_bls12381::ni_dkg::groth20_bls12_381::create_resharing_transcript

pub fn create_resharing_transcript(
    threshold: NumberOfNodes, 
    number_of_receivers: NumberOfNodes, 
    csp_dealings: &BTreeMap<NodeIndex, Dealing>, 
    resharing_public_coefficients: &PublicCoefficientsBytes
) -> Result<Transcript, CspDkgCreateReshareTranscriptError>

Creates an NiDKG transcript with the same public key as an existing threshold key.

Prerequisites

The dealings MUST be verified before calling this method; otherwise dealers may provide receivers with invalid shares or dealers may provide keys that do not preserve the public key.

Errors

CspDkgCreateReshareTranscriptError::InsufficientDealingsError if the length of csp_dealings is smaller than the new threshold (set by resharing_public_coefficients).

CspDkgCreateReshareTranscriptError::ResharingFailed if the public key created by this resharing does not match the previous public key.

CspDkgCreateTranscriptError::InvalidThresholdError if the threshold is either zero, or larger than the number_of_receivers.

CspDkgCreateTranscriptError::InvalidDealingError if csp_dealings is malformed or invalid.

Panics

If Langrange interpolation fails (because of duplicate x-coordinates).