Function ic_crypto_internal_threshold_sig_bls12381::ni_dkg::groth20_bls12_381::update_forward_secure_epoch

pub fn update_forward_secure_epoch(
    secret_key: &FsEncryptionSecretKey, 
    epoch: Epoch, 
    seed: Randomness
) -> FsEncryptionSecretKey

Forgets keys before the given epoch.

Note: If the lowest supported epoch of the secret key is greater than or equal to the threshold cutoff epoch, the secret key is unchanged.

Arguments

secret_key - The forward-secure encryption key to be updated.

epoch - The earliest epoch at which to retain keys.

seed - Randomness used in updating the secret key to the given epoch.

Returns

A new FsEncryptionSecretKey, with keys only at or after the given epoch.