Function ic_crypto_internal_threshold_sig_bls12381::ni_dkg::groth20_bls12_381::verify_dealing

pub fn verify_dealing(
    _dkg_id: NiDkgId, 
    dealer_index: NodeIndex, 
    threshold: NumberOfNodes, 
    epoch: Epoch, 
    receiver_keys: &BTreeMap<NodeIndex, FsEncryptionPublicKey>, 
    dealing: &Dealing
) -> Result<(), CspDkgVerifyDealingError>

Verifies a dealing.

Arguments

dealer_index - The index of the dealer that provided the given dealing.

threshold - The threshold required by the given dealing.

epoch - The forward-secure encryption epoch used to encrypt the receivers' shares.

receiver_keys - The forward-secure encryption public keys used to encrypt the receivers' shares.

dealing - The dealing to verify.

Errors

CspDkgVerifyDealingError::InvalidThresholdError if the threshold is less than 1.

CspDkgVerifyDealingError::InvalidThresholdError if the threshold is greater than the number of receivers.

MisnumberedReceiverError if the receiver indices are not 0..num_receivers-1 inclusive.

CspDkgVerifyDealingError::InvalidDealingError if

The share indices are not 0..num_receivers-1 inclusive.

Any shares are malformed.

CspDkgVerifyDealingError::InvalidDealingError if dealing.public_coefficients.len() != threshold.

CspDkgVerifyDealingError::MalformedDealingError if any component of dealing is malformed or invalid.

CspDkgVerifyDealingError::InvalidDealingError if any one of the decryptability or sharing proofs of the dealing, or the integrity of the dealing ciphertexts, don't verify.