Function ic_crypto_internal_threshold_sig_bls12381::ni_dkg::groth20_bls12_381::verify_forward_secure_key

pub fn verify_forward_secure_key(
    public_key: &FsEncryptionPublicKey, 
    pop: &FsEncryptionPop, 
    associated_data: &[u8]
) -> Result<(), ()>

Verifies that a public key is a point on the curve and that the proof of possession holds.

Errors

Err(()) if

Any of the components of public_key is not a correct group element.

The proof of possession doesn't verify.