Function ic_crypto_internal_threshold_sig_bls12381::ni_dkg::groth20_bls12_381::verify_resharing_dealing

pub fn verify_resharing_dealing(
    dkg_id: NiDkgId, 
    dealer_resharing_index: NodeIndex, 
    threshold: NumberOfNodes, 
    epoch: Epoch, 
    receiver_keys: &BTreeMap<NodeIndex, FsEncryptionPublicKey>, 
    dealing: &Dealing, 
    resharing_public_coefficients: &PublicCoefficientsBytes
) -> Result<(), CspDkgVerifyDealingError>

Verifies a dealing created to reshare an existing secret.

Also cf. verify_dealing.

Arguments

dkg_id - Unused.

dealer_resharing_index - The index of the dealer that provided the given dealing.

threshold - The threshold required by the given dealing.

epoch - The forward-secure encryption epoch used to encrypt the receivers' shares.

receiver_keys - The forward-secure encryption public keys used to encrypt the receivers' shares.

dealing - The dealing to verify.

resharing_public_coefficients - The PublicCoefficients corresponding to the secret share that is being re-shared.

Errors

Same errors as verify_dealing.

CspDkgVerifyDealingError::InvalidDealingError if the constant term in the public coefficient doesn't equal the individual public key of the dealer in the resharing instance.

CspDkgVerifyDealingError::InvalidDealingError if the dealer's public key is malformed.

Panics

If there are no public_coefficients in dealing.