Function ic_crypto_internal_threshold_sig_bls12381::ni_dkg::groth20_bls12_381::types::convert_keyset_to_keyset_with_pop

pub fn convert_keyset_to_keyset_with_pop(
    key_set: FsEncryptionKeySet
) -> FsEncryptionKeySetWithPop

Converts an old FsEncryptionKeySet to a FsEncryptionKeySetWithPop.

The old FsEncrptionPok is formatted as a FsEncryptionPop as follows:

The blinder of the PoK is written as the pop_key value in the PoP

The challenge of the proof of possession is set equal to 0.

The response of the Pok is written as the response value in the PoP,

Security Notice

The reformatted PoK does not constitute a valid PoP. This function must be used for compatibility purposes only and it will be removed as part of CRP-923.