Function ic_crypto_internal_threshold_sig_bls12381::types::public_coefficients::conversions::pub_key_bytes_from_pub_coeff_bytes

pub fn pub_key_bytes_from_pub_coeff_bytes(
    public_coefficients: &PublicCoefficientsBytes
) -> PublicKeyBytes

Returns the public key associated to the given public_coefficients.

Empty public_coefficients represent an all-zero polynomial, so in this case we return a zero-public key.